Refactor configuration modules
