+++ docs/ru/architecture/ARCHITECTURE.md
@@ -72,7 +72,7 @@ flowchart LR
 | `library/pipelines/*` | Чтение входных CSV, вызовы API, объединение данных, валидация и экспорт. |
 | `library/qa`, `library/table_quality.py` | Проверки схем, подсчёт метрик качества, предупреждения. |
 | `library/postprocessing` | Упорядочивание колонок, генерация метаданных, сортировка. |
-| `library/config.py` | Слияние YAML, переменных окружения и CLI, проверка типов. |
+| `library/config/` | Слияние YAML, переменных окружения и CLI, проверка типов. |
 
 ## Модель выполнения
