+++ library/config/__init__.py
@@ -1,0 +1,137 @@
+"""Public configuration API consolidating models, loader, and runtime helpers."""
+
+from __future__ import annotations
+
+from .models import (
+    ActivityActionTypeCfg,
+    ActivityBoundsCfg,
+    ActivityCfg,
+    ActivityEnrichmentCfg,
+    ActivityPropertiesCfg,
+    ApiCfg,
+    AssayCfg,
+    ChemblCacheCfg,
+    ChemblSourceCfg,
+    Config,
+    ConfigError,
+    ConfigMetadata,
+    ConfigSource,
+    CrossRefCfg,
+    DocQualityCfg,
+    DocTypeCfg,
+    DocumentAllCfg,
+    DocumentCfg,
+    DocumentChemblCfg,
+    DocumentPubmedCfg,
+    IupharCfg,
+    InitCfg,
+    IoCfg,
+    LocalCfg,
+    LogCfg,
+    MoleculeCatalogCfg,
+    OpenAlexCfg,
+    PubChemCfg,
+    PubMedCfg,
+    RateCfg,
+    ResourcesCfg,
+    RetryCfg,
+    SemanticScholarCfg,
+    SourcesCfg,
+    SystemCfg,
+    TargetAllCfg,
+    TargetCfg,
+    TargetChemblCfg,
+    TargetChemblBatchRetryCfg,
+    TargetIupharCfg,
+    TargetUniprotCfg,
+    TestitemBatchRetryCfg,
+    TestitemCfg,
+    TestitemMoleculeEnrichmentCfg,
+    TestitemMoleculeEnrichmentFlagsCfg,
+    TestitemMoleculeEnrichmentLoggingCfg,
+    TestitemMoleculeEnrichmentOutputCfg,
+    TestitemMoleculeEnrichmentSourcesCfg,
+    TESTITEM_FIELD_DEFAULTS,
+    UniprotCfg,
+    UniprotMappingCfg,
+    build_alias_map,
+)
+from .loader import (
+    ensure_dirs,
+    load_config,
+    print_config,
+    _absolutise_path_value,
+    _mask_secrets,
+    _serialize_paths,
+)
+from .runtime import (
+    configure_rate_limiters,
+    crossref_session,
+    openalex_session,
+    session_with_retry,
+)
+
+__all__ = [
+    "ActivityActionTypeCfg",
+    "ActivityBoundsCfg",
+    "ActivityCfg",
+    "ActivityEnrichmentCfg",
+    "ActivityPropertiesCfg",
+    "ApiCfg",
+    "AssayCfg",
+    "ChemblCacheCfg",
+    "ChemblSourceCfg",
+    "Config",
+    "ConfigError",
+    "ConfigMetadata",
+    "ConfigSource",
+    "CrossRefCfg",
+    "DocQualityCfg",
+    "DocTypeCfg",
+    "DocumentAllCfg",
+    "DocumentCfg",
+    "DocumentChemblCfg",
+    "DocumentPubmedCfg",
+    "IupharCfg",
+    "InitCfg",
+    "IoCfg",
+    "LocalCfg",
+    "LogCfg",
+    "MoleculeCatalogCfg",
+    "OpenAlexCfg",
+    "PubChemCfg",
+    "PubMedCfg",
+    "RateCfg",
+    "ResourcesCfg",
+    "RetryCfg",
+    "SemanticScholarCfg",
+    "SourcesCfg",
+    "SystemCfg",
+    "TargetAllCfg",
+    "TargetCfg",
+    "TargetChemblCfg",
+    "TargetChemblBatchRetryCfg",
+    "TargetIupharCfg",
+    "TargetUniprotCfg",
+    "TestitemBatchRetryCfg",
+    "TestitemCfg",
+    "TestitemMoleculeEnrichmentCfg",
+    "TestitemMoleculeEnrichmentFlagsCfg",
+    "TestitemMoleculeEnrichmentLoggingCfg",
+    "TestitemMoleculeEnrichmentOutputCfg",
+    "TestitemMoleculeEnrichmentSourcesCfg",
+    "TESTITEM_FIELD_DEFAULTS",
+    "UniprotCfg",
+    "UniprotMappingCfg",
+    "build_alias_map",
+    "configure_rate_limiters",
+    "crossref_session",
+    "ensure_dirs",
+    "load_config",
+    "openalex_session",
+    "print_config",
+    "session_with_retry",
+    "_absolutise_path_value",
+    "_mask_secrets",
+    "_serialize_paths",
+]
